changed ip to localhost changed names of classes and methods
