+++ src/test/java/CreatePublicGroupTest.java
@@ -1,23 +1,24 @@
 import org.testng.annotations.Test;
-import pages.AdvertisingPage;
 import pages.LoginPage;
 import pages.MainPage;
+import pages.PublicCreatePage;
 
 import static org.testng.Assert.assertEquals;
 import static utils.PropertiesClass.getPropertyValue;
 
-public class AdvertisingCreateTest extends BaseTest {
+public class CreatePublicGroupTest extends BaseTest{
 
     @Test
-    public void advertisingCreateTest() {
+    public void createPublicGroupTest(){
         LoginPage loginPage = new LoginPage(driver);
         MainPage mainPage = new MainPage(driver);
-        AdvertisingPage advertisingPage = new AdvertisingPage(driver);
+        PublicCreatePage publicCreatePage = new PublicCreatePage(driver);
 
         loginPage.open();
         loginPage.doLogin(getPropertyValue("login"), getPropertyValue("password"));
         assertEquals(mainPage.findUserNavigationLabel().isDisplayed(), true);
-        advertisingPage.open();
-        assertEquals(advertisingPage.isAdsManagerButtonAvailable(), true);
+        publicCreatePage.open();
+        assertEquals(publicCreatePage.publicCreatPageAvailable(), true);
     }
+
 }
